sr: Only use overlay for MDP 5.x and cleanup a bit
The overlay code currently tries to parse the MDP version to support
MDP 4.0 and higher.  The parsing code fails on most (all?) MDP 4.x
devices because it expects the version to have three digits (4xx).
But the version reported by the kernel only has two digits.  This
means that in practice overlay is only used for MDP 5.x.

Enabling overlay for MDP 4.x revealed the following issues:

 * Older devices with MDP 4.1 and the kernel "mini ION upgrade" such
   as p930 are unable to mmap() the ION memory properly.

 * Newer devices with MDP 4.4 and full ION support such as mako do
   work, but the minui graphics code would need adjusted to work
   because it does not handle the alpha channel properly.

The easiest solution to these issues is to just continue using the
fbdev backend for MDP 4.x.

Also cleanup a couple of error paths while we're here:

 * mmap failure is indicated by MAP_FAILED, not NULL.

 * return NULL when overlay_init fails.

Change-Id: I5236f601fd967111fea736b7382a2295c0551260